Resolve circular dependecy by moving the common functions to common package
